Purging of cache is disabled by default
Enable purging of outdated cache files on each request. If you don't purge the cache yourself (i.e. using crontab and the method PurgeCache), you should change this to true. This feature was introduced because if 2 methods were used for purging, a rare condition could occur when the crontab deleted a cached file at the smae time when the cache file was about tom be read.